+++ core/src/js/components/secondary-nav/accordion/events/OnClickAccordion.js
@@ -1,4 +1,4 @@
-import EventAbstract from '../../nav/events/EventAbstract';
+import EventAbstract from '../../common/events/EventAbstract';
 
 /**
  * ActivePath Class
