+++ core/src/js/components/secondary-nav/common/events/OnArrowRight.js
