+++ core/src/js/components/secondary-nav/common/events/OnArrowUp.js
