+++ core/src/js/components/secondary-nav/common/events/OnSpace.js
