+++ core/src/js/components/secondary-nav/common/events/OnTab.js
