+++ core/src/js/components/secondary-nav/common/globals.js
