+++ core/src/scss/components/secondary-nav/_secondary-nav.scss
@@ -70,7 +70,7 @@
 // First level items
 // ---------------------------------------------------------------------------
 .su-secondary-nav__menu-lv1 {
-
+  // Nothing to see here.
 }
 
 // Secondary Level Items
